Use the real file name for the kotlin poet builder
Kotlin poet apparently uses this name property as a unique id within
its systems while the file is being built. It doesn't have any real
bearing on the file that's written (we pass a stream to the api for
writing), but keeping the names consistent in the kotlin poet system
and the file that's going to be written to disk seems like a good idea.

The docs (https://square.github.io/kotlinpoet/) suggest typically the
file name is passed without the extension.